Creating extensions for attributes and initialiser for vhdl types into MetaMachine types.
