+++ Sources/MetaMachines/VHDLMachines/VHDLMachinesConverter.swift
@@ -46,226 +46,11 @@ struct VHDLMachinesConverter {
             semantics: .vhdl,
             name: machine.name,
             initialState: machine.states[machine.initialState].name,
-            states: machine.states.map { toState(state: $0, machine: machine) },
+            states: machine.states.map { State(vhdl: $0, in: machine) },
             dependencies: [],
-            attributes: machineAttributes(machine: machine),
+            attributes: machine.attributes,
             metaData: []
         )
-    }
-    
-    private func arrangementAttributes(arrangement: VHDLMachines.Arrangement) -> [AttributeGroup] {
-        var attributes: [AttributeGroup] = []
-        let variables = AttributeGroup(
-            name: "variables",
-            fields: [
-                Field(name: "clocks", type: .table(columns: [
-                    ("name", .line),
-                    ("frequency", .integer),
-                    ("unit", .enumerated(validValues: Set(VHDLMachines.Clock.FrequencyUnit.allCases.map { $0.rawValue })))
-                ])),
-                Field(name: "external_signals", type: .table(columns: [
-                    ("mode", .enumerated(validValues: Set(VHDLMachines.Mode.allCases.map { $0.rawValue }))),
-                    ("type", .expression(language: .vhdl)),
-                    ("name", .line),
-                    ("comment", .line)
-                ])),
-                Field(name: "external_variables", type: .table(columns: [
-                    ("type", .expression(language: .vhdl)),
-                    ("name", .line),
-                    ("comment", .line)
-                ]))
-            ],
-            attributes: [
-                "clocks": .table(
-                    arrangement.clocks.map(\.toLineAttribute),
-                    columns: [
-                        ("name", .line),
-                        ("frequency", .integer),
-                        ("unit", .enumerated(validValues: Set(VHDLMachines.Clock.FrequencyUnit.allCases.map { $0.rawValue })))
-                    ]
-                ),
-                "external_signals": .table(
-                    arrangement.externalSignals.map(\.toLineAttribute),
-                    columns: [
-                        ("mode", .enumerated(validValues: Set(VHDLMachines.Mode.allCases.map { $0.rawValue }))),
-                        ("type", .expression(language: .vhdl)),
-                        ("name", .line),
-                        ("comment", .line)
-                    ]
-                ),
-                "external_variables": .table(
-                    arrangement.externalVariables.map(\.toLineAttribute),
-                    columns: [
-                        ("type", .expression(language: .vhdl)),
-                        ("name", .line),
-                        ("comment", .line)
-                    ]
-                )
-            ],
-            metaData: [:]
-        )
-        attributes.append(variables)
-        return attributes
-    }
-    
-    private func machineAttributes(machine: VHDLMachines.Machine) -> [AttributeGroup] {
-        var attributes: [AttributeGroup] = []
-        let variableFields: [Field] = [
-            Field(name: "clocks", type: .table(columns: [
-                ("name", .line),
-                ("frequency", .integer),
-                ("unit", .enumerated(validValues: Set(VHDLMachines.Clock.FrequencyUnit.allCases.map { $0.rawValue })))
-            ])),
-            Field(name: "external_signals", type: .table(columns: [
-                ("mode", .enumerated(validValues: Set(VHDLMachines.Mode.allCases.map { $0.rawValue }))),
-                ("type", .expression(language: .vhdl)),
-                ("name", .line),
-                ("value", .expression(language: .vhdl)),
-                ("comment", .line)
-            ])),
-            Field(name: "generics", type: .table(columns: [
-                ("type", .expression(language: .vhdl)),
-                ("name", .line),
-                ("value", .expression(language: .vhdl)),
-                ("comment", .line)
-            ])),
-            Field(name: "machine_signals", type: .table(columns: [
-                ("type", .expression(language: .vhdl)),
-                ("name", .line),
-                ("value", .expression(language: .vhdl)),
-                ("comment", .line)
-            ])),
-            Field(name: "machine_variables", type: .table(columns: [
-                ("type", .expression(language: .vhdl)),
-                ("name", .line),
-                ("value", .expression(language: .vhdl)),
-                ("comment", .line)
-            ])),
-            Field(name: "driving_clock", type: .enumerated(validValues: Set(machine.clocks.map { $0.name })))
-        ]
-        let variableAttributes: [String: Attribute] = [
-            "clocks": .table(
-                machine.clocks.map(\.toLineAttribute),
-                columns: [
-                    ("name", .line),
-                    ("frequency", .integer),
-                    ("unit", .enumerated(validValues: Set(VHDLMachines.Clock.FrequencyUnit.allCases.map { $0.rawValue })))
-                ]
-            ),
-            "external_signals": .table(
-                machine.externalSignals.map(\.toLineAttribute),
-                columns: [
-                    ("mode", .enumerated(validValues: Set(VHDLMachines.Mode.allCases.map { $0.rawValue }))),
-                    ("type", .expression(language: .vhdl)),
-                    ("name", .line),
-                    ("value", .expression(language: .vhdl)),
-                    ("comment", .line)
-                ]
-            ),
-            "generics": .table(
-                machine.generics.map(\.toLineAttribute),
-                columns: [
-                    ("type", .expression(language: .vhdl)),
-                    ("name", .line),
-                    ("value", .expression(language: .vhdl)),
-                    ("comment", .line)
-                ]
-            ),
-            "machine_signals": .table(
-                machine.machineSignals.map(\.toLineAttribute),
-                columns: [
-                    ("type", .expression(language: .vhdl)),
-                    ("name", .line),
-                    ("value", .expression(language: .vhdl)),
-                    ("comment", .line)
-                ]
-            ),
-            "machine_variables": .table(
-                machine.machineVariables.map(\.toLineAttribute),
-                columns: [
-                    ("type", .expression(language: .vhdl)),
-                    ("name", .line),
-                    ("value", .expression(language: .vhdl)),
-                    ("comment", .line)
-                ]
-            ),
-            "driving_clock": .enumerated(machine.clocks[machine.drivingClock].name, validValues: Set(machine.clocks.map { $0.name }))
-        ]
-        let variables = AttributeGroup(
-            name: "variables",
-            fields: variableFields,
-            attributes: variableAttributes,
-            metaData: [:]
-        )
-        attributes.append(variables)
-        let parameters = AttributeGroup(
-            name: "parameters",
-            fields: [
-                Field(name: "is_parameterised", type: .bool),
-                Field(name: "parameter_signals", type: .table(columns: [
-                    ("type", .expression(language: .vhdl)),
-                    ("name", .line),
-                    ("value", .expression(language: .vhdl)),
-                    ("comment", .line)
-                ])),
-                Field(name: "returnable_signals", type: .table(columns: [
-                    ("type", .expression(language: .vhdl)),
-                    ("name", .line),
-                    ("comment", .line)
-                ]))
-            ],
-            attributes: [
-                "is_parameterised": .bool(machine.isParameterised),
-                "parameter_signals": .table(
-                    !machine.isParameterised ? [] : machine.parameterSignals.map(\.toLineAttribute),
-                    columns: [
-                        ("type", .expression(language: .vhdl)),
-                        ("name", .line),
-                        ("value", .expression(language: .vhdl)),
-                        ("comment", .line)
-                    ]
-                ),
-                "returnable_signals": .table(
-                    !machine.isParameterised ? [] : machine.returnableSignals.map(\.toLineAttribute),
-                    columns: [
-                        ("type", .expression(language: .vhdl)),
-                        ("name", .line),
-                        ("comment", .line)
-                    ]
-                )
-            ],
-            metaData: [:]
-        )
-        attributes.append(parameters)
-        let includes = AttributeGroup(
-            name: "includes",
-            fields: [
-                Field(name: "includes", type: .code(language: .vhdl)),
-                Field(name: "architecture_head", type: .code(language: .vhdl)),
-                Field(name: "architecture_body", type: .code(language: .vhdl))
-            ],
-            attributes: [
-                "includes": .code(machine.includes.reduce("", addNewline), language: .vhdl),
-                "architecture_head": .code(machine.architectureHead ?? "", language: .vhdl),
-                "architecture_body": .code(machine.architectureBody ?? "", language: .vhdl)
-            ],
-            metaData: [:]
-        )
-        attributes.append(includes)
-        let settings = AttributeGroup(
-            name: "settings",
-            fields: [
-                Field(name: "initial_state", type: .enumerated(validValues: Set([""] + machine.states.map(\.name)))),
-                Field(name: "suspended_state", type: .enumerated(validValues: Set([""] + machine.states.map(\.name))))
-            ],
-            attributes: [
-                "initial_state": .enumerated(machine.states[machine.initialState].name, validValues: Set(machine.states.map(\.name) + [""])),
-                "suspended_state": .enumerated(machine.suspendedState.map { machine.states[$0].name } ?? "", validValues: Set([""] + machine.states.map(\.name)))
-            ],
-            metaData: [:]
-        )
-        attributes.append(settings)
-        return attributes
     }
     
     private func addNewline(lhs: String, rhs: String) -> String {
@@ -277,110 +62,9 @@ struct VHDLMachinesConverter {
         }
         return lhs + "\n" + rhs
     }
-    
-    private func stateAttributes(state: VHDLMachines.State, machine: VHDLMachines.Machine) -> [AttributeGroup] {
-        var attributes: [AttributeGroup] = []
-        let externals = machine.externalSignals.map { $0.name }
-        let variables = AttributeGroup(
-            name: "variables",
-            fields: [
-                Field(name: "externals", type: .enumerableCollection(
-                        validValues: Set(externals)
-                )),
-                Field(name: "state_signals", type: .table(columns: [
-                    ("type", .expression(language: .vhdl)),
-                    ("name", .line),
-                    ("value", .expression(language: .vhdl)),
-                    ("comment", .line)
-                ])),
-                Field(name: "state_variables", type: .table(columns: [
-                    ("type", .expression(language: .vhdl)),
-                    ("lower_range", .line),
-                    ("upper_range", .line),
-                    ("name", .line),
-                    ("value", .expression(language: .vhdl)),
-                    ("comment", .line)
-                ]))
-            ],
-            attributes: [
-                "externals": .enumerableCollection(Set(state.externalVariables), validValues: Set(externals)),
-                "state_signals": .table(
-                    state.signals.map(\.toLineAttribute),
-                    columns: [
-                        ("type", .expression(language: .vhdl)),
-                        ("name", .line),
-                        ("value", .expression(language: .vhdl)),
-                        ("comment", .line)
-                    ]
-                ),
-                "state_variables": .table(
-                    state.variables.map(\.toLineAttribute),
-                    columns: [
-                        ("type", .expression(language: .vhdl)),
-                        ("lower_range", .line),
-                        ("upper_range", .line),
-                        ("name", .line),
-                        ("value", .expression(language: .vhdl)),
-                        ("comment", .line)
-                    ]
-                )
-            ],
-            metaData: [:]
-        )
-        attributes.append(variables)
-        let order = AttributeGroup(
-            name: "actions",
-            fields: [
-                Field(name: "action_names", type: .table(columns: [
-                    ("name", .line)
-                ])),
-                Field(name: "action_order", type: .table(columns: [
-                    ("timeslot", .integer),
-                    ("action", .enumerated(validValues: Set(state.actions.keys)))
-                ]))
-            ],
-            attributes: [
-                "action_names": .table(state.actions.keys.sorted().map { [LineAttribute.line($0)] }, columns: [
-                    ("name", .line)
-                ]),
-                "action_order": .table(state.actionOrder.toLineAttribute(validValues: Set(state.actions.keys)), columns: [
-                    ("timeslot", .integer),
-                    ("action", .enumerated(validValues: Set(state.actions.keys)))
-                ])
-            ],
-            metaData: [:]
-        )
-        attributes.append(order)
-        return attributes
-    }
-
-    private func toState(state: VHDLMachines.State, machine: VHDLMachines.Machine) -> State {
-        let actions = state.actionOrder.reduce([]){ $0 + $1 }.map {
-            toAction(actionName: $0, code: state.actions[$0] ?? "")
-        }
-        guard let stateIndex = machine.states.firstIndex(where: { $0.name == state.name }) else {
-            fatalError("Cannot find state with name: \(state.name).")
-        }
-        return State(
-            name: state.name,
-            actions: actions,
-            transitions: machine.transitions.filter({ $0.source == stateIndex }).map({ toTransition(transition: $0, machine: machine) }),
-            attributes: stateAttributes(state: state, machine: machine),
-            metaData: []
-        )
-    }
 
     private func toAction(actionName: String, code: String) -> Action {
         Action(name: actionName, implementation: code, language: .vhdl)
-    }
-
-    private func toTransition(transition: VHDLMachines.Transition, machine: VHDLMachines.Machine) -> Transition {
-        Transition(
-            condition: transition.condition,
-            target: machine.states[transition.target].name,
-            attributes: [],
-            metaData: []
-        )
     }
